pyproject.toml
[project]
name = "cubari-hakken"
version = "0.1.0"
description = "Cubari.moe/Guya.moe sources indexer and searcher"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.0",
    "requests>=2.32.5",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
]
